系统用户返回id修改为long类型

https://codeup.aliyun.com/5f765d05728df4b180fab723/framework/merge_request/1131674

+++ springcloud-tester/com.longyou.commservice/src/main/java/com/longyou/comm/admin/service/IUserAdminService.java
@@ -3,5 +3,5 @@ package com.longyou.comm.admin.service;
 import org.cloud.model.TFrameUser;
 
 public interface IUserAdminService {
-    public int saveOrUpdate(TFrameUser tFrameUser) throws Exception;
+    public long saveOrUpdate(TFrameUser tFrameUser) throws Exception;
 }
